* Updating the gitignore. * Updating snyk to remember to fix the current issues in 30 days. * Adding travis. * Adding coveralls. * Adding/Updating Akamai CopyRight Header to all files.
