+++ .gitignore
@@ -1,3 +1,5 @@
 .idea
 node_modules/
 releases/
+.nyc_output
+coverage
