+++ .travis.yml
@@ -1,0 +1,5 @@
+language: node_js
+node_js:
+  - 8.0
+script: npm test
+after_success: npm run coveralls
