+++ package.json
@@ -11,8 +11,9 @@
     "lint": "eslint --ignore-path .gitignore .",
     "start": "node bin/akamaiAuth",
     "pretest": "npm run lint",
-    "test": "snyk test && mocha test --timeout 0",
-    "build": "node release.js"
+    "test": "snyk test && nyc --reporter=html --reporter=text mocha test --timeout 0",
+    "build": "node release.js",
+    "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
   "main": "./index.js",
   "files": [
@@ -20,6 +21,7 @@
     "src"
   ],
   "dependencies": {
+    "akamai-edge-grid-auth": "^1.0.0",
     "akamai-node-release": "^1.0.1",
     "chalk": "^2.3.0",
     "child-process-promise": "^2.2.1",
@@ -39,9 +41,11 @@
   "devDependencies": {
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
+    "coveralls": "^3.0.1",
     "eslint": "^4.19.1",
     "eslint-config-loopback": "^10.0.0",
     "mocha": "^5.1.1",
+    "nyc": "^11.7.3",
     "snyk": "^1.80.0"
   }
 }
